Remove embed-page-specific Webpack entry.
