--- client/stylesheets/_embedded.scss
@@ -1,13 +1,0 @@
-/** @format */
-
-// Import our wp-admin reset.
-@import './shared/_reset.scss';
-
-// Import any global styles.
-@import './shared/_global.scss';
-
-// Import the embed-specific styles.
-@import './shared/_embed.scss';
-
-// Import gutenberg SCSS so that we can apply our theme colors.
-@import './shared/gutenberg-components.scss';
