+++ client/stylesheets/_index.scss
@@ -27,5 +27,8 @@
 // Import any global styles.
 @import './shared/_global.scss';
 
+// Import the embed-specific styles.
+@import './shared/_embed.scss';
+
 // Import gutenberg SCSS so that we can apply our theme colors.
 @import './shared/gutenberg-components.scss';
